[Issue #10760] Modify LoadOracleData to capture timing metrics for all parts of its run

+++ api/src/task/base_task.py
@@ -77,7 +77,7 @@ class BaseTask(abc.ABC, metaclass=abc.ABCMeta):
     def set_metrics(self, metrics: dict[str, Any]) -> None:
         self.metrics.update(**metrics)
 
-    def increment(self, name: str, value: int = 1, prefix: str | None = None) -> None:
+    def increment(self, name: str, value: float = 1, prefix: str | None = None) -> None:
         if name not in self.metrics:
             self.metrics[name] = 0
